refactor(guardian): extract remaining duplication across flat-hook hosts
Two more sources of new-code duplication SonarCloud still flagged after
the first pass (10.0%, threshold 3%):

- amazonq-guardian-operations.js duplicated kiro-guardian-operations.js's
  operation-list body almost verbatim (both flat {hooks: {event:
  [{matcher, command}]}} hosts). Extracted the shared shape into
  guardian-flat-hook-operations.js; both host-specific files are now thin
  config wrappers around it.
- The three adapters' buildGuardianInput bodies differed only in the
  shell tool name and which field carries the working directory.
  Extracted createBashToolGuardianInputMapper() into adapter-stdin-
  json.js, parameterized by those two values.

Kiro's own tests (13/13) and the full suite (3694/3694) still pass
unchanged -- both refactors are behavior-preserving.
