+++ scripts/hooks/goose-guardian-adapter.js
@@ -31,25 +31,9 @@
 'use strict';
 
 const { run } = require('./pre-bash-guardian-validate');
-const { bootstrapPlainExitCodeAdapter } = require('../lib/adapter-stdin-json');
+const { bootstrapPlainExitCodeAdapter, createBashToolGuardianInputMapper } = require('../lib/adapter-stdin-json');
 
-function buildGuardianInput(event) {
-  if (!event || typeof event !== 'object') {
-    return null;
-  }
-  if (event.tool_name !== 'developer__shell') {
-    return null;
-  }
-  const command = event.tool_input?.command;
-  if (!command || typeof command !== 'string') {
-    return null;
-  }
-  const input = { tool_name: 'Bash', tool_input: { command } };
-  if (typeof event.working_dir === 'string') {
-    input.cwd = event.working_dir;
-  }
-  return input;
-}
+const buildGuardianInput = createBashToolGuardianInputMapper({ shellToolName: 'developer__shell', cwdKey: 'working_dir' });
 
 module.exports = bootstrapPlainExitCodeAdapter({
   isMain: require.main === module,
