+++ scripts/hooks/openhands-guardian-adapter.js
@@ -23,25 +23,9 @@
 'use strict';
 
 const { run } = require('./pre-bash-guardian-validate');
-const { bootstrapPlainExitCodeAdapter } = require('../lib/adapter-stdin-json');
+const { bootstrapPlainExitCodeAdapter, createBashToolGuardianInputMapper } = require('../lib/adapter-stdin-json');
 
-function buildGuardianInput(event) {
-  if (!event || typeof event !== 'object') {
-    return null;
-  }
-  if (event.tool_name !== 'terminal') {
-    return null;
-  }
-  const command = event.tool_input?.command;
-  if (!command || typeof command !== 'string') {
-    return null;
-  }
-  const input = { tool_name: 'Bash', tool_input: { command } };
-  if (typeof event.working_dir === 'string') {
-    input.cwd = event.working_dir;
-  }
-  return input;
-}
+const buildGuardianInput = createBashToolGuardianInputMapper({ shellToolName: 'terminal', cwdKey: 'working_dir' });
 
 module.exports = bootstrapPlainExitCodeAdapter({
   isMain: require.main === module,
