+++ scripts/lib/adapter-stdin-json.js
@@ -103,9 +103,37 @@ function bootstrapPlainExitCodeAdapter({ isMain, buildGuardianInput, runGuardian
   return { buildGuardianInput };
 }
 
+// Amazon Q, Goose, and OpenHands all deliver {tool_name, tool_input:
+// {command}, <cwdKey>} on stdin for their shell tool, differing only in
+// the shell tool's name and which top-level field carries the working
+// directory -- this builds the translation function each adapter exports
+// as buildGuardianInput, instead of every host repeating the same four
+// guard checks with different literals.
+function createBashToolGuardianInputMapper({ shellToolName, cwdKey }) {
+  return function buildGuardianInput(event) {
+    if (!event || typeof event !== 'object') {
+      return null;
+    }
+    if (event.tool_name !== shellToolName) {
+      return null;
+    }
+    const command = event.tool_input?.command;
+    if (!command || typeof command !== 'string') {
+      return null;
+    }
+    const input = { tool_name: 'Bash', tool_input: { command } };
+    const cwdValue = event[cwdKey];
+    if (typeof cwdValue === 'string') {
+      input.cwd = cwdValue;
+    }
+    return input;
+  };
+}
+
 module.exports = {
   MAX_STDIN,
   bootstrapPlainExitCodeAdapter,
+  createBashToolGuardianInputMapper,
   readAdapterStdinJson,
   runPlainExitCodeGuardianAdapter,
 };
